Implementation of specificNaming and a simple JUnit test.
This allows more configuration for the service when deployed
via the command line. Instead of naming processed files by timestamp,
the service can be set to grab info from a the message using an
XPath Expression and use that as an identifying label. This means
multiple updates to the same record will result in only one, up-to-date
file for ingesting.